requirements.txt
ffmpeg-python==0.2.0
setuptools==78.1.0
openai-whisper==20231117
tqdm==4.66.1
numpy==1.26.4
torch==2.2.0 
python-dotenv==1.0.1 
chromadb==0.6.3
langchain_ollama==0.3.0
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.10.6
pydantic_settings==2.8.1